requirements.txt
Flask
requests
python-dotenv
flask_caching
pendulum